+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 5.1
 - Install add-ons in InstallDatabase.php
 - Rename "Edit Student Grades" program to "Historical Grades" in Menu.php, Help_en.php & help.po
 - CSS Allow for browser address bar to disappear on scroll in zresponsive.css
+- JS Prevent scrolling body while Menu is open in warehouse.js
 
 Changes in 5.0.5
 ----------------